+++ android-app/app/src/main/java/com/lorenzo/mangadownloader/VyMangaSource.kt
@@ -37,6 +37,7 @@ class VyMangaSource(
         id = MangaSourceIds.VYMANGA,
         displayName = "VyManga",
         shortName = "VY",
+        language = MangaSourceLanguage.ENG,
     )
 
     override val invalidChapterUrlMessage: String =